move follow button into its own component
